Fixed issue with extra events for virtual particles
We had the issues with the extra events on the walls like y=A-1 for
virtual particles which are not in the system: we shouldn't calculate
such events at all and this path fix this issue. The code desn't work
yet, we have new strange error in create_virt_particle function.